Restructure jax_aiter into two-layer architecture (ops/ + recipes)
Layer 1 (jax_aiter/ops/): Single-kernel FFI wrappers with no custom_vjp
or custom_partitioning. Explicit parameters, no env var reads. Reusable
building blocks for custom training recipes.

  - ops/gemm_fp4.py: gemm_fp4(), cast_mxfp4(), cast_mxfp4_dual()
  - ops/gemm_bf16.py: gemm_bf16()
  - ops/mha.py: mha_fwd(), mha_bwd(), MhaFwdConfig, MhaBwdConfig
  - ops/rmsnorm.py: rmsnorm_fwd()
  - ops/activation.py: silu_and_mul()

Layer 2 (existing modules): Training recipes with custom_vjp +
custom_partitioning that use ops/ internally. All existing public APIs
unchanged.

Also: remove dead mha/context_parallel.py placeholder.

Test: 364 passed, 5 xfailed (pre-existing MHA d>=96).
