+++ jax_aiter/gemm/gemm.py
@@ -31,43 +31,7 @@ import jax.numpy as jnp
 from jax.experimental.custom_partitioning import custom_partitioning
 from jax.sharding import NamedSharding, PartitionSpec as P
 
-from ..ffi.registry import register_ffi_target
-
-
-def _ensure_registered():
-    register_ffi_target("GemmFwdJA", "ROCM")
-
-
-def _gemm_fwd_call(out_shape, sem_shape, dtype):
-    call = jax.ffi.ffi_call(
-        "GemmFwdJA",
-        (
-            jax.ShapeDtypeStruct(out_shape, dtype),
-            jax.ShapeDtypeStruct(sem_shape, jnp.uint32),
-        ),
-        vmap_method="broadcast_all",
-        input_layouts=[None, None],
-        output_layouts=[None, None],
-        has_side_effect=False,
-    )
-
-    def _invoke(a, b):
-        out, _ = call(a, b)
-        return out
-
-    return jax.jit(_invoke)
-
-
-# ---------------------------------------------------------------------------
-# Raw (unpartitioned) GEMM -- used inside custom_partitioning lowering.
-# ---------------------------------------------------------------------------
-def _gemm_raw(a, b):
-    """Raw FFI GEMM call. A[M,K] @ B[N,K]^T -> Out[M,N]."""
-    _ensure_registered()
-    M, K = a.shape
-    N = b.shape[0]
-    fn = _gemm_fwd_call((M, N), (16, 64), a.dtype)
-    return fn(a, b)
+from ..ops.gemm_bf16 import gemm_bf16 as _gemm_raw
 
 
 # ---------------------------------------------------------------------------
